WIP Keyboard events on text inputs: extract out component + fixes
 Now implemented in adding new annotations and editing existing
 annotations
