+++ src/sidebar-overlay/comment-box/components/comment-box-form.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react'
-import ReactDOM from 'react-dom'
 import { ClickHandler } from '../../types'
-import * as constants from '../constants'
 import TagsContainer from './tag-input-container'
 import { Tooltip } from 'src/common-ui/components'
 import { getLocalStorage } from 'src/util/storage'
 import { TAG_SUGGESTIONS_KEY } from 'src/constants'
 import cx from 'classnames'
+import TextInputControlled from 'src/common-ui/components/TextInputControlled'
+import TextAreaDynamicRows from 'src/common-ui/components/TextAreaDynamicRows'
 
 const styles = require('./comment-box-form.css')
 
@@ -138,19 +138,16 @@ class CommentBoxForm extends React.Component<Props, State> {
 
         return (
             <React.Fragment>
-                <TextArea
-                    // todo: re-implement what is needed here:
-                    // todo: value might become default value, check expandRows is actually used, take out textRef if not used, see if tagInput is used.
-                    // expandRows={true}
-                    // value={commentText}
-                    // onClick={() => {
-                    //     this.setTagInputActive(false)
-                    //     this.setState(state => ({ showTagsPicker: false }))
-                    // }}
-                    specialHandlers={[this.onEnterSaveHandler]}
+                <TextInputControlled
+                    defaultValue={commentText}
+                    onClick={() => {
+                        this.setTagInputActive(false)
+                        this.setState(state => ({ showTagsPicker: false }))
+                    }}
                     className={styles.textArea}
                     placeholder="Add a private note... (save with cmd/ctrl+enter)"
                     onChange={this.props.handleCommentTextChange}
+                    specialHandlers={[this.onEnterSaveHandler]}
                 />
 
                 {/* Save and Cancel buttons. */}
@@ -203,331 +200,3 @@ class CommentBoxForm extends React.Component<Props, State> {
 }
 
 export default CommentBoxForm
-
-// This Component exists to work-around issues with keypress events in our extension
-// due to the way it's loaded in a page's DOM as react and ShadowDOM,
-// and perhaps aggravated by the react-retarget work-around package too.
-
-// Hopefully when React itself refactors it's event handling code, this temporary work-around will become obsolete
-// ( https://github.com/facebook/react/issues/13525 & https://github.com/facebook/react/issues/15257 )
-
-// The issue we face is that when typing into an un-controlled react text input in our extension
-// (e.g. sidebar annotation comment), key-presses that should be exclusively characters input to the field,
-// get instead picked up by keypress event handlers on that page, e.g:
-// - the '?' key on github & twitter bringing up help.
-// - the 't' key on our extensions shortcuts itself opening up the tagging.
-
-// We work around this here by explicitly handling all key-presses needed for text editing on an input
-// and manually acting on their intent, (e.g. just inserting a 't' character, or deleting a selection)
-// then we specifically set the event to not continue propagating to other event handlers.
-// These events are only listening when the component is mounted. When the component is not mounted
-// a page's key handlers will continue to fire normally.
-
-// Events that fire from the react retarget package are given a path array containing the tree
-// of HTML elements that this event should come from. (real event is on shadow root, but path will
-// indicate the intended element)
-type ReTargetedTextElementEvent = React.KeyboardEvent<HTMLTextAreaElement> & {
-    path: (HTMLElement & HTMLTextAreaElement)[]
-}
-
-interface TextAreaProps {
-    onChange: (s: string) => void
-    specialHandlers?: { test: (e) => boolean; handle: (e) => void }[]
-}
-interface TextAreaState {
-    content: string
-    selection: { startOffset: number; endOffset: number }
-}
-class TextArea extends React.Component<
-    TextAreaProps & Partial<ReactDOM.IntrinsicElements.textarea>,
-    TextAreaState
-> {
-    textarea: HTMLTextAreaElement
-
-    static defaultProps = {
-        specialHandlers: [],
-    }
-
-    constructor(props) {
-        super(props)
-        this.state = {
-            content: '',
-            selection: { startOffset: 0, endOffset: 0 },
-        }
-    }
-
-    // -- Methods primarily to do with keeping the selection state in sync --
-    componentDidMount() {
-        this.registerEventListeners()
-        this.textarea.focus()
-    }
-
-    componentWillUnmount() {
-        this.deRegisterEventListeners()
-    }
-
-    registerEventListeners = () => {
-        this.textarea.addEventListener('select', this.selectHandler)
-        this.textarea.addEventListener('click', this.clickHandler)
-        this.textarea.addEventListener('focus', this.focusHandler)
-        this.textarea.addEventListener('keyup', this.keyupHandler)
-    }
-
-    deRegisterEventListeners = () => {
-        this.textarea.removeEventListener('select', this.selectHandler)
-        this.textarea.removeEventListener('click', this.clickHandler)
-        this.textarea.removeEventListener('focus', this.focusHandler)
-        this.textarea.removeEventListener('keyup', this.keyupHandler)
-    }
-
-    // We keep the selection (the highlighted characters in an input box) in sync
-    // with the components state, otherwise when react re-renders the component
-    // the selection will be reset to the end of the text.
-    selectHandler = () => this.updateSelectionState()
-    clickHandler = () => this.updateSelectionState()
-    focusHandler = () => this.updateSelectionState()
-    keyupHandler = () => this.updateSelectionState()
-
-    // Update the internal state representation of the text input's selection
-    updateSelectionState = () =>
-        this.setState({ selection: this.getSelectionFromDom() })
-
-    // Get the selection from the HTML component
-    getSelectionFromDom = () => ({
-        startOffset: this.textarea.selectionStart,
-        endOffset: this.textarea.selectionEnd,
-    })
-
-    // Important to keep the content (internal state + parent component handler) and the selection (internal state)
-    // in sync when changes are made outside of our managed key presses, e.g. Copy/Paste.
-    handleOnChange = () => {
-        this.updateTextarea({
-            content: this.textarea.value,
-            selection: this.getSelectionFromDom(),
-        })
-    }
-
-    // Set the selection from our state to the HTML component
-    syncSelectionToDom = (textareaRef, selection) => {
-        textareaRef.selectionStart = selection.startOffset
-        textareaRef.selectionEnd = selection.endOffset
-    }
-
-    // -- Methods primarily to do with key presses --
-
-    // The main logic intercepting key-presses
-    private handleTextAreaKeyDown = (e: ReTargetedTextElementEvent) => {
-        // First check if we have been given a special handler to check for by the parent component
-        // (e.g. Ctrl+Enter to save)
-        for (const specialHandler of this.props.specialHandlers) {
-            if (specialHandler.test(e)) {
-                specialHandler.handle(e)
-                return TextArea.stopProp(e)
-            }
-        }
-
-        // Otherwise, if we class it as a control event we handle it as such
-        // Or otherwise, if we class it as an input event we handle it as such
-        // stopping the propagation if either do so.
-        if (this.handleControlEvent(e) || this.handleInputTextEvent(e)) {
-            return TextArea.stopProp(e)
-        }
-
-        // Otherwise, we haven't done anything with the input event and it should continue propagating
-        return false
-    }
-
-    // If the input event matches a control character we know how to handle, then handle it manually
-    private handleControlEvent(e: ReTargetedTextElementEvent) {
-        switch (e.key) {
-            case 'Enter':
-                this.handleInput('\n')
-                return true
-            case 'ArrowLeft':
-                this.moveSelection(-1)
-                return true
-            case 'ArrowRight':
-                this.moveSelection(+1)
-                return true
-            case 'ArrowUp':
-                // todo: ArrowUp
-                // moveSelectionUp(el);
-                return true
-            case 'ArrowDown':
-                // todo: ArrowDown
-                // moveSelectionUp(el);
-                return true
-            case 'End':
-                this.jumpSelection(this.textarea.value.length)
-                return true
-            case 'PageDown':
-                this.jumpSelection(this.textarea.value.length)
-                return true
-            case 'Home':
-                this.jumpSelection(0)
-                return true
-            case 'PageUp':
-                this.jumpSelection(0)
-                return true
-            case 'Backspace':
-                this.handleBackspaceInput()
-                return true
-            case 'Delete':
-                // Delete if pressed without a multi char selection deletes the next char
-                // Otherwise for multi char selection, functions same as backspace
-                this.state.selection.endOffset ===
-                this.state.selection.startOffset
-                    ? this.handleDeleteInput()
-                    : this.handleBackspaceInput()
-                return true
-            default:
-                return false
-        }
-    }
-
-    private handleInputTextEvent(e: ReTargetedTextElementEvent) {
-        // todo: can we change to detecting keypress rather than keydown and therefore get real charCodes
-        // todo: we might then be able to use the code below to detect them a bit more explicitly
-        // if (event.which != 0 && event.charCode != 0)
-        //      char= String.fromCharCode(event.which);
-
-        // Here we take advantage of the the e.key either being a single character descriptor like 'A','?','0',etc
-        // or a key description like 'Enter', 'Backspace', etc. Single chars are for input, others are not.
-        const printable = e.key.length <= 1
-        if (printable && !(e.ctrlKey || e.metaKey)) {
-            this.handleInput(e.key)
-            // if (this.props.expandRows) {
-            //     this.handleRows(e)
-            // }
-            return true
-        }
-        return false
-    }
-
-    // Stop the event from propagating any further
-    private static stopProp(e) {
-        e.preventDefault()
-        e.stopPropagation()
-        return true
-    }
-
-    // todo: there was previously something to do with extending the rows, though it seemed like maybe the styles were overriding it
-    // todo: this happened on key input and on scroll
-    private handleRows(e) {
-        // const comment = e.target.value
-        // const rows =
-        //     comment.length === 0
-        //         ? constants.NUM_DEFAULT_ROWS
-        //         : Math.max(this.state.rows, constants.NUM_MIN_ROWS)
-        //
-        // if (rows !== this.state.rows) {
-        //     this.setState({ rows })
-        // }
-    }
-
-    private handleInput(char) {
-        const selection = this.getSelectionFromDom()
-        const text = this.textarea.value
-
-        const textBeforeSelection = text.substring(0, selection.startOffset)
-        const textAfterSelection = text.substring(selection.endOffset)
-        const newText = textBeforeSelection + char + textAfterSelection
-
-        // Whether a single char has been inserted in a single cursor selection or a selection of multiple characters
-        // always move the offset forward by one to account for that added character.
-        selection.endOffset = selection.startOffset + 1
-        selection.startOffset = selection.startOffset + 1
-
-        this.updateTextarea({ content: newText, selection })
-        this.props.onChange(this.textarea.value)
-    }
-
-    private handleBackspaceInput() {
-        const selection = this.getSelectionFromDom()
-        const text = this.textarea.value
-
-        const textBeforeSelection = text.substring(0, selection.startOffset)
-        const textAfterSelection = text.substring(selection.endOffset)
-        const newText =
-            selection.startOffset === selection.endOffset
-                ? textBeforeSelection.substring(
-                      0,
-                      textBeforeSelection.length - 1,
-                  ) + textAfterSelection
-                : textBeforeSelection + textAfterSelection
-
-        // Deleting one character moves the selection index back one if it is not already at the start
-        if (
-            selection.startOffset === selection.endOffset &&
-            selection.startOffset !== 0
-        ) {
-            selection.startOffset--
-            selection.endOffset--
-        } else {
-            // Otherwise the highlighted characters are deleted and the index goes to where the first char was
-            selection.endOffset = selection.startOffset
-        }
-
-        this.updateTextarea({ content: newText, selection })
-        this.props.onChange(this.textarea.value)
-    }
-
-    private handleDeleteInput() {
-        const selection = this.getSelectionFromDom()
-        const text = this.textarea.value
-        const newText =
-            text.substring(0, selection.startOffset) +
-            text.substring(
-                Math.min(text.length, selection.endOffset + 1),
-                text.length,
-            )
-
-        this.updateTextarea({ content: newText, selection })
-        this.props.onChange(this.textarea.value)
-    }
-
-    // Helper method to update the selection by an increment/decrement
-    private moveSelection(move) {
-        const selection = this.state.selection
-        selection.endOffset += move
-        selection.startOffset += move
-        this.updateTextarea({ content: this.textarea.value, selection })
-    }
-
-    // Helper method to jump the selection to a specific position
-    private jumpSelection(index) {
-        const selection = this.state.selection
-        selection.endOffset = index
-        selection.startOffset = index
-        this.updateTextarea({ content: this.textarea.value, selection })
-    }
-
-    // Helper method to update our state with intended content or selection for the textarea
-    updateTextarea = ({ content, selection }) => {
-        const updatedContent = content || this.textarea.value
-        const updatedSelection = selection || this.getSelectionFromDom()
-        this.setState(
-            {
-                content: updatedContent,
-                selection: updatedSelection,
-            },
-            () => this.syncSelectionToDom(this.textarea, updatedSelection),
-        )
-        this.props.onChange(this.state.content)
-    }
-
-    render() {
-        const { onChange, specialHandlers, ...props } = this.props
-        return (
-            <textarea
-                ref={c => {
-                    this.textarea = c
-                }}
-                value={this.state.content}
-                onChange={this.handleOnChange}
-                onKeyDown={this.handleTextAreaKeyDown}
-                {...props}
-            />
-        )
-    }
-}
